requirements.txt
fastapi
uvicorn
langchain-openai
langgraph
pydantic
typing-extensions
python-jose[cryptography]